*FINAL FINAL PUSH*
To those who are grading this project: While we did not use pushes to track who is coding which sections and how long each person spent on it, as we found that we were pushing it enough times that that could have been distracting, we tracked that information in other ways.

For instance, we divided the tasks up about evenly beforehand: 

1) We (Ariella and Tenzin) both wrote and debugged/proofed the sudoku generator.

2) I (Ariella) wrote and debugged/proofed the hint generator and automatic solver.

3) Tenzin wrote and debugged/proofed the interactive play component (in which the user enters their guess for a particular cell and get feedback on whether that number is right, wrong, or already filled in within the grid). 

Moreover, we have time stamps tracking when each of us worked on each component (please see previous commits for this information). Finally, we tracked the hours manually: We each spent about five hours on the sudou generator (four hours writing and one hour debugging each), I spent about two hours on the hint generator (one and a half hours writing, half an hour debugging), and around three hours on the solver (two hours writing, an hour debugging), and Tenzin spent two hours on the interactive play component (one and a half hours writing, half an hour debugging).